+++ examples/msvc6/basic_shapes/basic_shapes.cpp
@@ -177,7 +177,7 @@ BOOL Cbasic_shapesApp::InitInstance() {
   double size = 1.0;
 
   // Create a nice little cube
-  createCube(object,size);
+  createCube(object,(float)size);
 
   // Move the object over some so the Phantom will not initially be
   // inside the object.
@@ -521,7 +521,7 @@ void Cbasic_shapesApp::toggle_haptics(int enable) {
 void createCube(cMesh *mesh, float edge) {
 
   // I define the cube's "radius" to be half the edge size
-  float radius = edge / 2.0;
+  float radius = edge / 2.0f;
   int n;
   int cur_index = 0;
   int start_index = 0;
